Cheat Code/ Home button
I added a cheat code when you press ‘c’ in the game. When you press
escape in the game, the game now exits. I added a home button right
next to the reset button. I also changed the background for the main
menu just to move the over under sign to the right more