Created a level with a monster
The code was more complicated than I thought so I didn't really
implement the monsters the way I wanted to. Right now, any level without
a monster just has the monster set at a negative x position. It's also
probably not possible to create multiple monsters in a level without
doing a bunch of other stuff so let's try to avoid that